requirements.txt
binance_trading_bot==1.0.0
ccxt==1.44.41
pandas==1.3.3
matplotlib==3.4.3
flask==2.0.1